+++ test/kura_coverage2_SUITE.erl
@@ -100,7 +100,6 @@ groups() ->
 %%----------------------------------------------------------------------
 
 init_per_suite(Config) ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_test_repo:start(),
     kura_sandbox:start(),