+++ test/kura_custom_type_tests.erl
@@ -129,7 +129,6 @@ integration_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     application:set_env(pg_types, uuid_format, string),
     kura_test_repo:start(),