+++ test/kura_integration_assoc_tests.erl
@@ -54,7 +54,6 @@ assoc_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_test_repo:start(),
     create_tables(),