+++ test/kura_integration_tests.erl
@@ -184,7 +184,6 @@ integration_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_test_repo:start(),
     {ok, _} = kura_test_repo:query(