+++ test/kura_migration_stress_tests.erl
@@ -24,7 +24,6 @@ migration_stress_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_stress_repo:start(),
     kura_migrator:ensure_schema_migrations(kura_stress_repo),