+++ test/kura_paginator_tests.erl
@@ -34,7 +34,6 @@ paginator_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_test_repo:start(),
     {ok, _} = kura_test_repo:query(