+++ test/kura_uuid_tests.erl
@@ -51,7 +51,6 @@ uuid_test_() ->
     end}.
 
 setup() ->
-    application:ensure_all_started(pgo),
     application:ensure_all_started(kura),
     kura_test_repo:start(),
     {ok, _} = kura_test_repo:query(